restructure the files for reaxff, add frozendict dependency, add ase dependency for testing
