+++ jax_md/reaxff/reaxff_helper.py
@@ -7,7 +7,7 @@ Author: Mehmet Cagri Kaymak
 import jax
 import jax.numpy as jnp
 import numpy as onp
-from jax_md.reaxff_forcefield import ForceField
+from jax_md.reaxff.reaxff_forcefield import ForceField
 from dataclasses import fields
 from jax import custom_jvp
 
@@ -31,410 +31,6 @@ def vectorized_cond(pred, true_fun, false_fun, operand):
   true_op = jnp.where(pred, operand, 0)
   false_op = jnp.where(pred, 0, operand)
   return jnp.where(pred, true_fun(true_op), false_fun(false_op))
-
-def calculate_bo_single(distance,
-                rob1, rob2, rob3,
-                ptp, pdp, popi, pdo, bop1, bop2,
-                cutoff):
-  '''
-  This function is needed to find the cutoff for the bonded interactions
-  during parameter optimization
-
-  to get the highest bor (bond order potential):
-      rob1:
-      bop2: group 2, line 2, col. 6
-      bop1: group 2, line 2, col. 5
-
-      rob2:
-      ptp: group 2, line 2, col. 3
-      pdp: group 2, line 2, col. 2
-
-      rob3:
-      popi: group 2, line 1, col. 7
-      pdo: group 2, line 1, col. 5
-  '''
-
-  rhulp = jnp.where(rob1 <= 0, 0, distance / rob1)
-  rh2 = rhulp ** bop2
-  ehulp = (1 + cutoff) * jnp.exp(bop1 * rh2)
-  ehulp = jnp.where(rob1 <= 0, 0.0, ehulp)
-
-  rhulp2 = jnp.where(rob2 == 0, 0, distance / rob2)
-  rh2p = rhulp2 ** ptp
-  ehulpp = jnp.exp(pdp * rh2p)
-  ehulpp = jnp.where(rob2 <= 0, 0.0, ehulpp)
-
-
-  rhulp3 = jnp.where(rob3 == 0, 0, distance / rob3)
-  rh2pp = rhulp3 ** popi
-  ehulppp = jnp.exp(pdo*rh2pp)
-  ehulppp = jnp.where(rob3 <= 0, 0.0, ehulppp)
-
-  #print(ehulp , ehulpp , ehulppp)
-  bor = ehulp + ehulpp + ehulppp
-
-
-  return bor # if < cutoff, will be ignored
-
-def find_limits(type1, type2, force_field, cutoff):
-  vect_bo_function = jax.jit(jax.vmap(calculate_bo_single,
-                          in_axes=(0,None,None,None,None,
-                                   None,None,None,
-                                   None,None,None)),backend='cpu')
-  rob1 = force_field.rob1[type1,type2] # select max, typical values (0.1, 2)
-  bop2 = force_field.bop2[type1,type2] # select min, typical values (1,10)
-  bop1 = force_field.bop1[type1,type2] # select max, typical values (-0.2, -0.01)
-
-  rob2 = force_field.rob2[type1,type2]
-  ptp = force_field.ptp[type1,type2]
-  pdp = force_field.pdp[type1,type2]
-
-  rob3 = force_field.rob3[type1,type2]
-  popi = force_field.popi[type1,type2]
-  pdo = force_field.pdo[type1,type2]
-
-  distance = jnp.linspace(0.0, 10, 1000)
-
-  res = vect_bo_function(distance,
-                  rob1, rob2, rob3,
-                  ptp, pdp, popi, pdo, bop1, bop2,
-                  cutoff)
-
-  ind = jnp.sum(res > cutoff)
-
-  return distance[ind]
-
-def find_all_cutoffs(force_field,cutoff,atom_indices, cutoff_dist=5.0):
-  lenn = len(atom_indices)
-  cutoff_dict = dict()
-  for i in range(lenn):
-    type_i = atom_indices[i]
-    for j in range(i,lenn):
-      type_j = atom_indices[j]
-      dist = find_limits(type_i,type_j, force_field, cutoff)
-      cutoff_dict[(type_i,type_j)] = dist
-      cutoff_dict[(type_j,type_i)] = dist
-
-      if dist > cutoff_dist and type_i != -1 and type_j!=-1: #-1 TYPE is for the filler atom
-        dist = round(dist,2)
-        print(f"[WARNING] between type {type_i} and type {type_j}" +
-              "the bond length could be greater than {CLOSE_NEIGH_CUTOFF} A! ({dist} A)")
-        cutoff_dict[(type_i,type_j)] = cutoff_dist
-        cutoff_dict[(type_j,type_i)] = cutoff_dist
-  return cutoff_dict
-
-def parse_and_save_force_field(old_ff_file, new_ff_file,force_field):
-  output = ""
-  f = open(old_ff_file, 'r')
-  line = f.readline()
-  output = output + line
-  header = line.strip()
-
-  line = f.readline()
-  output = output + line
-  num_params = int(line.strip().split()[0])
-  global_params = jnp.zeros(shape=(num_params,1), dtype=jnp.float64)
-  ff = force_field
-  for i in range(num_params):
-    line = f.readline()
-    line = list(line)
-    #-------------------------------------------------------------
-    if i == 0:
-      line[:10] = "{:10.4f}".format(ff.over_coord1[0])  #overcoord1
-    if i == 1:
-      line[:10] = "{:10.4f}".format(ff.over_coord2[0]) #overcoord2
-    #-------------------------------------------------------------
-
-    #-------------------------------------------------------------
-    if i == 3:
-      line[:10] = "{:10.4f}".format(ff.trip_stab4[0])  #trip_stab4
-    if i == 4:
-      line[:10] = "{:10.4f}".format(ff.trip_stab5[0]) #trip_stab5
-    if i == 7:
-      line[:10] = "{:10.4f}".format(ff.trip_stab8[0])  #trip_stab8
-    if i == 10:
-      line[:10] = "{:10.4f}".format(ff.trip_stab11[0]) #trip_stab11
-    #-------------------------------------------------------------
-    #valency related parameters
-    if i == 2:
-      line[:10] = "{:10.4f}".format(ff.val_par3[0])  #val_par3
-    if i == 14:
-      line[:10] = "{:10.4f}".format(ff.val_par15[0]) #val_par15
-    if i == 15:
-      line[:10] = "{:10.4f}".format(ff.par_16[0])  #par_16
-    if i == 16:
-      line[:10] = "{:10.4f}".format(ff.val_par17[0]) #val_par17
-    if i == 17:
-      line[:10] = "{:10.4f}".format(ff.val_par18[0])  #val_par18
-    if i == 19:
-      line[:10] = "{:10.4f}".format(ff.val_par20[0]) #val_par20
-    if i == 20:
-      line[:10] = "{:10.4f}".format(ff.val_par21[0])  #val_par21
-    if i == 30:
-      line[:10] = "{:10.4f}".format(ff.val_par31[0]) #val_par31
-    if i == 33:
-      line[:10] = "{:10.4f}".format(ff.val_par34[0])  #val_par34
-    if i == 38:
-      line[:10] = "{:10.4f}".format(ff.val_par39[0]) #val_par39
-    #-------------------------------------------------------------
-
-    #-------------------------------------------------------------
-    #over-under coord.
-    if i == 5:
-      line[:10] = "{:10.4f}".format(ff.par_6[0])  #par_6
-    if i == 6:
-      line[:10] = "{:10.4f}".format(ff.par_7[0]) #par_7
-    if i == 8:
-      line[:10] = "{:10.4f}".format(ff.par_9[0])  #par_9
-    if i == 9:
-      line[:10] = "{:10.4f}".format(ff.par_10[0]) #par_10
-    if i == 31:
-      line[:10] = "{:10.4f}".format(ff.par_32[0])  #par_32
-    if i == 32:
-      line[:10] = "{:10.4f}".format(ff.par_33[0]) #par_33
-
-    #-------------------------------------------------------------
-    #torsion
-    if i == 23:
-      line[:10] = "{:10.4f}".format(ff.par_24[0])  #par_24
-    if i == 24:
-      line[:10] = "{:10.4f}".format(ff.par_25[0]) #par_25
-    if i == 25:
-      line[:10] = "{:10.4f}".format(ff.par_26[0])  #par_26
-    if i == 27:
-      line[:10] = "{:10.4f}".format(ff.par_28[0]) #par_28
-
-    #-------------------------------------------------------------
-    # vdw
-    if i == 28:
-      line[:10] = "{:10.4f}".format(ff.vdw_shiedling[0]) #vdw_shiedling
-    output = output + ''.join(line)
-
-  line = f.readline()
-  output = output + line
-
-  num_atom_types = int(line.strip().split()[0])
-  # skip 3 lines of comment
-  output = output + f.readline()
-  output = output + f.readline()
-  output = output + f.readline()
-
-  atom_names = []
-  line_ctr = 0
-  for i in range(num_atom_types):
-    # first line
-    line = f.readline()
-    line = list(line)
-    line[3 + 9 * 0:3 + 9 * 1] = "{:9.4f}".format(ff.rat[i]) #rat - rob1
-    line[3 + 9 * 3:3 + 9 * 4] = "{:9.4f}".format(ff.rvdw[i]) #rvdw
-    line[3 + 9 * 4:3 + 9 * 5] = "{:9.4f}".format(ff.eps[i]) #eps
-    line[3 + 9 * 5:3 + 9 * 6] = "{:9.4f}".format(ff.gamma[i]) #gamma
-    line[3 + 9 * 6:3 + 9 * 7] = "{:9.4f}".format(ff.rapt[i]) #rapt - rob2
-    line[3 + 9 * 7:3 + 9 * 8] = "{:9.4f}".format(ff.stlp[i]) #stlp
-
-    output = output + ''.join(line)
-
-    # second line
-    line = f.readline()
-    line = list(line)
-    line[3 + 9 * 0:3 + 9 * 1] = "{:9.4f}".format(ff.alf[i]) #alf
-    line[3 + 9 * 1:3 + 9 * 2] = "{:9.4f}".format(ff.vop[i]) #vop
-    line[3 + 9 * 2:3 + 9 * 3] = "{:9.4f}".format(ff.valf[i]) #valf
-    line[3 + 9 * 3:3 + 9 * 4] = "{:9.4f}".format(ff.valp1[i]) #valp1
-    line[3 + 9 * 5:3 + 9 * 6] = "{:9.4f}".format(ff.electronegativity[i])
-    line[3 + 9 * 6:3 + 9 * 7] = "{:9.4f}".format(ff.idempotential[i])
-
-    output = output + ''.join(line)
-    # third line
-    line = f.readline()
-    line = list(line)
-    line[3 + 9 * 0:3 + 9 * 1] = "{:9.4f}".format(ff.vnq[i]) #vnq - rob3
-    line[3 + 9 * 1:3 + 9 * 2] = "{:9.4f}".format(ff.vlp1[i]) #vlp1
-    line[3 + 9 * 3:3 + 9 * 4] = "{:9.4f}".format(ff.bo131[i]) #bo131
-    line[3 + 9 * 4:3 + 9 * 5] = "{:9.4f}".format(ff.bo132[i]) #bo132
-    line[3 + 9 * 5:3 + 9 * 6] = "{:9.4f}".format(ff.bo133[i]) #bo133
-
-    output = output + ''.join(line)
-
-    # fourth line
-    line = f.readline()
-    line = list(line)
-    line[3 + 9 * 0:3 + 9 * 1] = "{:9.4f}".format(ff.vovun[i])
-    line[3 + 9 * 1:3 + 9 * 2] = "{:9.4f}".format(ff.vval1[i])
-    line[3 + 9 * 3:3 + 9 * 4] = "{:9.4f}".format(ff.vval3[i])
-    line[3 + 9 * 4:3 + 9 * 5] = "{:9.4f}".format(ff.vval4[i])
-
-    output = output + ''.join(line)
-
-
-
-  line = f.readline()  # num_bonds
-  output = output + line
-
-  line = line.strip()
-  num_bonds = int(line.split()[0])
-  output = output + f.readline() # skip next line (comment)
-  for _ in range(num_bonds):
-    # first line
-    line = f.readline()
-    tmp = line.strip().split()
-    line = list(line)
-    i = int(tmp[0]) - 1 # index starts at 0
-    j = int(tmp[1]) - 1
-
-    line[6 + 9 * 0:6 + 9 * 1] = "{:9.4f}".format(ff.de1[i,j])
-    line[6 + 9 * 1:6 + 9 * 2] = "{:9.4f}".format(ff.de2[i,j])
-    line[6 + 9 * 2:6 + 9 * 3] = "{:9.4f}".format(ff.de3[i,j])
-    line[6 + 9 * 3:6 + 9 * 4] = "{:9.4f}".format(ff.psi[i,j])
-    line[6 + 9 * 4:6 + 9 * 5] = "{:9.4f}".format(ff.pdo[i,j])
-    line[6 + 9 * 5:6 + 9 * 6] = "{:9.4f}".format(ff.v13cor[i,j])
-    line[6 + 9 * 6:6 + 9 * 7] = "{:9.4f}".format(ff.popi[i,j])
-    line[6 + 9 * 7:6 + 9 * 8] = "{:9.4f}".format(ff.vover[i,j])
-    #print(''.join(line))
-    output = output + ''.join(line)
-    # second line
-    line = f.readline()
-    line = list(line)
-
-    line[6 + 9 * 0:6 + 9 * 1] = "{:9.4f}".format(ff.psp[i,j])
-    line[6 + 9 * 1:6 + 9 * 2] = "{:9.4f}".format(ff.pdp[i,j])
-    line[6 + 9 * 2:6 + 9 * 3] = "{:9.4f}".format(ff.ptp[i,j])
-    line[6 + 9 * 4:6 + 9 * 5] = "{:9.4f}".format(ff.bop1[i,j])
-    line[6 + 9 * 5:6 + 9 * 6] = "{:9.4f}".format(ff.bop2[i,j])
-    line[6 + 9 * 6:6 + 9 * 7] = "{:9.4f}".format(ff.ovc[i,j])
-    #print(''.join(line))
-    output = output + ''.join(line)
-  line = f.readline()  # num_off_diag
-  output = output + line
-
-  line = line.strip()
-  num_off_diag = int(line.split()[0])
-
-  for _ in range(num_off_diag):
-    # first line
-    # first line
-    line = f.readline()
-    tmp = line.strip().split()
-    line = list(line)
-    i = int(tmp[0]) - 1 # index starts at 0
-    j = int(tmp[1]) - 1
-
-    line[6 + 9 * 0:6 + 9 * 1] = "{:9.4f}".format(ff.p2co_off[i,j])
-    line[6 + 9 * 1:6 + 9 * 2] = "{:9.4f}".format(ff.p1co_off[i,j])  # was /2
-    line[6 + 9 * 2:6 + 9 * 3] = "{:9.4f}".format(ff.p3co_off[i,j])
-    line[6 + 9 * 3:6 + 9 * 4] = "{:9.4f}".format(ff.rob1_off[i,j])
-    line[6 + 9 * 4:6 + 9 * 5] = "{:9.4f}".format(ff.rob2_off[i,j])
-    line[6 + 9 * 5:6 + 9 * 6] = "{:9.4f}".format(ff.rob3_off[i,j])
-
-    output = output + ''.join(line)
-
-  #valency angle parameters
-  line = f.readline()  # num_val_params
-  output = output + line
-
-  line = line.strip()
-  num_val_params = int(line.split()[0])
-
-  for _ in range(num_val_params):
-    # first line
-    line = f.readline()
-    tmp = line.strip().split()
-    line = list(line)
-    i = int(tmp[0]) - 1 # index starts at 0
-    j = int(tmp[1]) - 1
-    k = int(tmp[2]) - 1
-
-    line[9 + 9 * 0:9 + 9 * 1] = "{:9.4f}".format(ff.th0[i,j,k])
-    line[9 + 9 * 1:9 + 9 * 2] = "{:9.4f}".format(ff.vka[i,j,k])
-    line[9 + 9 * 2:9 + 9 * 3] = "{:9.4f}".format(ff.vka3[i,j,k])
-    line[9 + 9 * 3:9 + 9 * 4] = "{:9.4f}".format(ff.vka8[i,j,k])
-    line[9 + 9 * 4:9 + 9 * 5] = "{:9.4f}".format(ff.vkac[i,j,k])
-    line[9 + 9 * 5:9 + 9 * 6] = "{:9.4f}".format(ff.vkap[i,j,k])
-    line[9 + 9 * 6:9 + 9 * 7] = "{:9.4f}".format(ff.vval2[i,j,k])
-
-    output = output + ''.join(line)
-
-  #torsion parameters
-  line = f.readline()  # num_tors_params
-  output = output + line
-
-  line = line.strip()
-  num_tors_params = int(line.split()[0])
-
-  for _ in range(num_tors_params):
-    # first line
-    line = f.readline()
-    tmp = line.strip().split()
-    line = list(line)
-    i1 = int(tmp[0]) - 1 # index starts at 0
-    i2 = int(tmp[1]) - 1
-    i3 = int(tmp[2]) - 1
-    i4 = int(tmp[3]) - 1
-
-    if i1 != -1 and i4 != -1:
-      line[12 + 9 * 0:12 + 9 * 1] = "{:9.4f}".format(ff.v1[i1,i2,i3,i4])
-      line[12 + 9 * 1:12 + 9 * 2] = "{:9.4f}".format(ff.v2[i1,i2,i3,i4])
-      line[12 + 9 * 2:12 + 9 * 3] = "{:9.4f}".format(ff.v3[i1,i2,i3,i4])
-      line[12 + 9 * 3:12 + 9 * 4] = "{:9.4f}".format(ff.v4[i1,i2,i3,i4])
-      line[12 + 9 * 4:12 + 9 * 5] = "{:9.4f}".format(ff.vconj[i1,i2,i3,i4])
-
-    if i1 == -1 and i4 == -1:
-      sel_ind = force_field.num_atom_types - 1
-      line[12 + 9 * 0:12 + 9 * 1] = "{:9.4f}".format(ff.v1[sel_ind,
-                                                           i2,
-                                                           i3,
-                                                           sel_ind])
-      line[12 + 9 * 1:12 + 9 * 2] = "{:9.4f}".format(ff.v2[sel_ind,
-                                                           i2,
-                                                           i3,
-                                                           sel_ind])
-      line[12 + 9 * 2:12 + 9 * 3] = "{:9.4f}".format(ff.v3[sel_ind,
-                                                           i2,
-                                                           i3,
-                                                           sel_ind])
-      line[12 + 9 * 3:12 + 9 * 4] = "{:9.4f}".format(ff.v4[sel_ind,
-                                                           i2,
-                                                           i3,
-                                                           sel_ind])
-      line[12 + 9 * 4:12 + 9 * 5] = "{:9.4f}".format(ff.vconj[sel_ind,
-                                                              i2,
-                                                              i3,
-                                                              sel_ind])
-    output = output + ''.join(line)
-
-  # hbond parameters
-  #torsion parameters
-  line = f.readline()  # num_tors_params
-  output = output + line
-
-  line = line.strip()
-  num_hbond_params = int(line.split()[0])
-
-  for i in range(num_hbond_params):
-    line = f.readline()
-    tmp = line.strip().split()
-    line = list(line)
-    i1 = int(tmp[0]) - 1
-    i2 = int(tmp[1]) - 1
-    i3 = int(tmp[2]) -1
-    line[9 + 9 * 0:9 + 9 * 1] = "{:9.4f}".format(ff.rhb[i1,i2,i3])
-    line[9 + 9 * 1:9 + 9 * 2] = "{:9.4f}".format(ff.dehb[i1,i2,i3])
-    line[9 + 9 * 2:9 + 9 * 3] = "{:9.4f}".format(ff.vhb1[i1,i2,i3])
-    line[9 + 9 * 3:9 + 9 * 4] = "{:9.4f}".format(ff.vhb2[i1,i2,i3])
-    output = output + ''.join(line)
-
-
-
-  # need to append some extra lines because of 0 values
-  for line in f:
-    output = output + line
-
-  file_new = open(new_ff_file,"w")
-  file_new.write(output)
-  file_new.close()
-
-  f.close()
-
 
 def init_params_for_filler_atom_type(FF_field_dict):
   #TODO: make sure that index -1 doesnt belong to a real atom!!!
