+++ jax_md/reaxff/reaxff_interactions.py
@@ -4,7 +4,8 @@ Contains interaction list related functions for ReaxFF
 Author: Mehmet Cagri Kaymak
 """
 from jax_md import space, partition, util
-from jax_md.reaxff_energy import calculate_bo, calculate_reaxff_energy
+from jax_md.reaxff.reaxff_energy import (calculate_bo, 
+                                        calculate_reaxff_energy)
 from typing import Callable, Any, Tuple
 import jax
 import jax.numpy as jnp
